lara/common: offer revert to pistols in TR2
Resolves #3455.
