+++ src/libtrx/config/map.def
@@ -35,6 +35,7 @@ X_CFG_BOOL(gameplay.fix_item_duplication_glitch,     true)
 X_CFG_BOOL(gameplay.fix_qwop_glitch,                 false)
 X_CFG_BOOL(gameplay.fix_wade_wall_hit,               true)
 X_CFG_BOOL(gameplay.fix_walk_run_jump,               true)
+X_CFG_BOOL(gameplay.revert_to_pistols,               false)
 X_CFG_BOOL(input.enable_responsive_passport,         true)
 X_CFG_BOOL(input.enable_tr3_sidesteps,               true)
 X_CFG_BOOL(profile.new_game_plus_unlock,             false)
