+++ src/libtrx/config/map_tr1.def
@@ -37,7 +37,6 @@ X_CFG_BOOL(gameplay.fix_step_glitch,           true)
 X_CFG_BOOL(gameplay.fix_water_exit,            false)
 X_CFG_ENUM(gameplay.jump_lock_mode,            JUMP_LOCK_TUNED, JUMP_LOCK_MODE)
 X_CFG_BOOL(gameplay.restore_ps1_enemies,       false)
-X_CFG_BOOL(gameplay.revert_to_pistols,         false)
 X_CFG_BOOL(input.enable_buffering,             false)
 X_CFG_BOOL(input.enable_numeric_keys,          true)
 X_CFG_BOOL(rendering.enable_vsync,             true)
