+++ src/libtrx/include/libtrx/game/lara/common.h
@@ -9,6 +9,7 @@ LARA_INFO *Lara_GetLaraInfo(void);
 ITEM *Lara_GetItem(void);
 void Lara_Initialise(const GF_LEVEL *level);
 extern void Lara_InitialiseInventory(const GF_LEVEL *level);
+void Lara_RevertToPistolsIfNeeded(void);
 void Lara_SetStartAnimState(LARA_EXTRA_STATE state);
 bool Lara_IsControllable(void);
 void Lara_SetControllable(bool controllable);
