+++ src/tr1/game/lara/common.h
@@ -9,5 +9,3 @@
 void Lara_InitialiseLoad(int16_t item_num);
 
 void Lara_UseItem(GAME_OBJECT_ID obj_id);
-
-void Lara_RevertToPistolsIfNeeded(void);
